requirements.txt
# Youth Green Jobs & Waste Recycling Hub - Backend Dependencies
# Django and REST Framework
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.10

# Environment and Configuration
python-decouple==3.8
python-dotenv==1.0.1
setuptools==75.1.0

# Image Processing
Pillow==11.3.0

# Additional packages for production
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==3.0.1
django-redis==5.4.0

# Additional packages added during setup
requests==2.32.3
django-filter==24.3
python-dateutil==2.8.2

# Development and Testing (optional)
# pytest-django==4.8.0
# coverage==7.4.0
# flake8==7.0.0
